Se ajusta putPutProfileUser para actualizar información de perfil del usuario
