+++ src/pages/Platform/PanelUser/00NavBar/04Configuration/01Profile/styles.module.css
@@ -1,15 +1,18 @@
 .container {
-    padding: 30px 0;
+    height: calc(100vh - 70px);
+    width: calc(100vw - 350px);
 }
-.containerCard {
-    width: 550px;
+.container__Component {
+    height: calc(100vh - 70px);
+    width: calc(100vw - 350px);
 }
 .title {
     color: #f28705;
-    font-size: 50px;
     font-weight: 600;
     letter-spacing: 1px;
-    font-family: 'Roboto', sans-serif;
+}
+.container__Card {
+    width: 550px;
 }
 .containerImageIcon {
     border: 1px solid #adb5bd;
@@ -75,6 +78,7 @@
 }
 .label {
     margin-bottom: 5px;
+    font-weight: 600;
 }
 .containerInput {
     width: 100%;
