+++ src/pages/Platform/PanelUser/00NavBar/04Configuration/05Mailconfiguration/UseYourEmail.tsx
