+++ src/pages/Platform/PanelUser/00NavBar/04Configuration/styles.module.css
@@ -7,8 +7,6 @@
     width: calc(100vw - 350px);
 }
 
-
-
 /*---------------UserRendering---------------*/
 .containerBranchPage {
     min-height: calc(100vh - 120px);
@@ -18,13 +16,6 @@
     padding: 10px 20px;
     border: 1px solid rgba(0, 0, 0, 0.2);
     cursor: pointer;
-}
-.title {
-    color: #f28705;
-    font-size: 50px;
-    font-weight: 600;
-    letter-spacing: 1px;
-    font-family: 'Roboto', sans-serif;
 }
 .active {
     background: #718bd8;
@@ -40,11 +31,4 @@
 }
 .containerUserCard {
     width: 550px;
-}
-.title {
-    color: #f28705;
-    font-size: 50px;
-    font-weight: 600;
-    letter-spacing: 1px;
-    font-family: 'Roboto', sans-serif;
 }
