+++ src/pages/Platform/PanelUser/03Inventories/01InventoryAssets/ConsultAssets/styles.module.css
@@ -10,7 +10,6 @@
     color: #f28705;
     font-weight: 600;
     letter-spacing: 1px;
-    font-family: 'Century Gothic', sans-serif;
 }
 .link__Head_Navigate {
     color: #fff;
